# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: The Champions Club API
  version: 1.0.0
  description: "GENERATED from tools/api/endpoints.mjs — DO NOT EDIT BY HAND. Human-readable version: API_CONTRACT.md. Schemas are derived from shared/types/*.ts (TypeScript is authoritative; inline/complex shapes may appear as free-form)."
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        success:
          description: "TS: false"
        error:
          type: object
          properties:
            code:
              description: "TS: ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
          required:
            - code
            - message
      required:
        - success
        - error
...
    ShopOrderDetail:
      allOf:
        - $ref: "#/components/schemas/ShopOrder"
        - type: object
          properties:
            subtotal:
              type: string
              pattern: ^\d+\.\d{2}$
              example: "1250.00"
            total_amount:
              type: string
              pattern: ^\d+\.\d{2}$
              example: "1250.00"
            amount_paid:
              type: string
              pattern: ^\d+\.\d{2}$
              example: "1250.00"
            payment_status:
              type: string
              enum:
                - PENDING
                - PARTIALLY_PAID
                - PAID
                - REFUNDED
                - NOT_REQUIRED
            items:
              type: array
              items:
                $ref: "#/components/schemas/ShopOrderItem"
            member_name:
              type: string
              nullable: true
            member_code:
              type: string
              nullable: true
          required:
            - subtotal
            - total_amount
            - amount_paid
            - payment_status
            - items
            - member_name
            - member_code
...
paths:
...
  /shop/orders:
    post:
      operationId: shop.orderCreate
      tags:
        - Shop
      summary: "Place a shop order: member online order (pickup/delivery) or staff counter sale."
      description: |-
        IN_STORE = a counter sale made by staff; PICKUP / DELIVERY = placed online by a member (no separate channel column: it follows from fulfillment). One transaction: `UPDATE products SET stock_quantity = stock_quantity - :q WHERE id = :id AND stock_quantity >= :q` per line (0 rows => OUT_OF_STOCK, rollback) + snapshot unit_price + discount_amount = subtotal x shop_discount_percent + delivery_fee (unless free_delivery_above) + payment. Totals and payment_status are derived (view shop_order_totals). IN_STORE orders are created COMPLETED; online orders start PLACED.

        Requirements: FR-SHOP-003, FR-SHOP-004, FR-SHOP-005, FR-SHOP-006, FR-SHOP-007, FR-SHOP-008, FR-INV-001, FR-FIN-001
      x-roles:
        - MEMBER
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-SHOP-003
        - FR-SHOP-004
        - FR-SHOP-005
        - FR-SHOP-006
        - FR-SHOP-007
        - FR-SHOP-008
        - FR-INV-001
        - FR-FIN-001
      x-tables:
        - shop_orders
        - shop_order_items
        - products
        - payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        description: 1-99
                    required:
                      - product_id
                      - quantity
                  description: 1-50 lines; duplicate products are merged
                fulfillment:
                  type: string
                  enum:
                    - PICKUP
                    - DELIVERY
                    - IN_STORE
                  description: "MEMBER: PICKUP|DELIVERY. Staff: IN_STORE"
                delivery_address:
                  type: string
                  description: required when fulfillment = DELIVERY
                member_id:
                  type: string
                  format: uuid
                  description: "Staff only: book/order on behalf of this member. MEMBER callers: omit (implied = self)."
                guest_name:
                  type: string
                  description: Walk-in name. Staff only. Required when no member_id.
                guest_phone:
                  type: string
                  description: Walk-in phone (E.164 or 10-digit). Staff only.
                payment_method:
                  type: string
                  enum:
                    - CASH
                    - CARD
                    - UPI
                    - ONLINE
                  description: "MEMBER: ONLINE. Staff counter sale: CASH|CARD|UPI (required for IN_STORE)"
              required:
                - items
                - fulfillment
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/ShopOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "402":
          description: PAYMENT_FAILED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: PRODUCT_NOT_FOUND | MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: OUT_OF_STOCK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: DELIVERY_ADDRESS_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: shop.orderList
      tags:
        - Shop
      summary: List shop orders (members see their own).
      description: "Requirements: FR-SHOP-011, FR-SHOP-009"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-SHOP-011
        - FR-SHOP-009
      x-tables:
        - shop_orders
        - shop_order_items
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - PLACED
              - CONFIRMED
              - READY_FOR_PICKUP
              - OUT_FOR_DELIVERY
              - COMPLETED
              - CANCELLED
        - name: fulfillment
          in: query
          required: false
          schema:
            type: string
            enum:
              - PICKUP
              - DELIVERY
              - IN_STORE
        - name: member_id
          in: query
          required: false
          description: staff only
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShopOrderDetail"
                  message:
                    type: string
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /shop/orders/{id}:
    get:
      operationId: shop.orderGet
      tags:
        - Shop
      summary: Order detail with lines.
      description: "Requirements: FR-SHOP-011, FR-SHOP-009"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-SHOP-011
        - FR-SHOP-009
      x-tables:
        - shop_orders
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/ShopOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /shop/orders/{id}/cancel:
    post:
      operationId: shop.orderCancel
      tags:
        - Shop
      summary: "Cancel an order: restores stock and refunds."
      description: |-
        MEMBER may cancel only own order while PLACED. Staff while not COMPLETED/CANCELLED. Adds the quantities back to products.stock_quantity and refunds the payment in full (the order then reports payment_status REFUNDED).

        Requirements: FR-SHOP-010, FR-INV-001
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-SHOP-010
        - FR-INV-001
      x-tables:
        - shop_orders
        - products
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/ShopOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: INVALID_STATUS_TRANSITION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /bar/menu:
    get:
      operationId: bar.menu
      tags:
        - Cafe
      summary: Bar & cafeteria menu.
      description: "Requirements: FR-BAR-001, FR-PUB-005"
      x-roles:
        - PUBLIC
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-001
        - FR-PUB-005
      x-tables:
        - bar_menu_items
      security: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum:
              - FOOD
              - SNACK
              - DRINK
        - name: include_unavailable
          in: query
          required: false
          description: staff only
          schema:
            type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BarMenuItem"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bar/menu-items:
    post:
      operationId: bar.menuCreate
      tags:
        - Cafe
      summary: Add a menu item.
      description: "Requirements: FR-BAR-011"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-011
      x-tables:
        - bar_menu_items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: unique
                category:
                  type: string
                  enum:
                    - FOOD
                    - SNACK
                    - DRINK
                description:
                  type: string
                price:
                  type: string
                  pattern: ^\d+\.\d{2}$
                  description: tax-inclusive
                sort_order:
                  type: integer
              required:
                - name
                - category
                - price
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BarMenuItem"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bar/menu-items/{id}:
    patch:
      operationId: bar.menuUpdate
      tags:
        - Cafe
      summary: Edit a menu item / mark unavailable.
      description: "Requirements: FR-BAR-011"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-011
      x-tables:
        - bar_menu_items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  type: string
                  enum:
                    - FOOD
                    - SNACK
                    - DRINK
                description:
                  type: string
                price:
                  type: string
                  pattern: ^\d+\.\d{2}$
                is_available:
                  type: boolean
                sort_order:
                  type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BarMenuItem"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: MENU_ITEM_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bar/orders:
    post:
      operationId: bar.orderCreate
      tags:
        - Cafe
      summary: Take a cafe order for a table label, member or guest. Appears instantly on the kitchen board.
      description: |-
        At least one of table_label, member_id, guest_name. Member discount = subtotal x bar_discount_percent of the member's CURRENT plan, stored as discount_amount (automatic; never asked). Guests pay list price. Snapshot unit_price + item_name. The order starts NEW. Totals and payment_status are derived (view bar_order_totals).

        Requirements: FR-BAR-003, FR-BAR-004, FR-BAR-005, FR-BAR-008, FR-BAR-013, FR-KIT-001
      x-roles:
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-003
        - FR-BAR-004
        - FR-BAR-005
        - FR-BAR-008
        - FR-BAR-013
        - FR-KIT-001
      x-tables:
        - bar_orders
        - bar_order_items
        - payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table_label:
                  type: string
                  description: free text, e.g. "T3" or "Court side"
                member_id:
                  type: string
                  format: uuid
                  description: "Staff only: book/order on behalf of this member. MEMBER callers: omit (implied = self)."
                guest_name:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      bar_menu_item_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        description: 1-50
                      notes:
                        type: string
                    required:
                      - bar_menu_item_id
                      - quantity
                  description: 1-50 lines
                payment_method:
                  type: string
                  enum:
                    - CASH
                    - CARD
                    - UPI
                    - ONLINE
                  description: pay now (CASH|CARD|UPI); omit to pay later at the desk
                notes:
                  type: string
              required:
                - items
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BarOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "402":
          description: PAYMENT_FAILED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: MENU_ITEM_NOT_FOUND | MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: MENU_ITEM_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: bar.orderList
      tags:
        - Cafe
      summary: List cafe orders (members see their own).
      description: "Requirements: FR-BAR-012, FR-BAR-009"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-012
        - FR-BAR-009
      x-tables:
        - bar_orders
        - bar_order_items
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - NEW
              - PREPARING
              - READY
              - SERVED
              - CANCELLED
        - name: payment_status
          in: query
          required: false
          schema:
            type: string
            enum:
              - PENDING
              - PARTIALLY_PAID
              - PAID
              - REFUNDED
              - NOT_REQUIRED
        - name: table_label
          in: query
          required: false
          schema:
            type: string
        - name: member_id
          in: query
          required: false
          description: staff only
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BarOrderDetail"
                  message:
                    type: string
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bar/orders/{id}:
    get:
      operationId: bar.orderGet
      tags:
        - Cafe
      summary: Cafe order detail.
      description: "Requirements: FR-BAR-012"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-012
      x-tables:
        - bar_orders
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BarOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bar/orders/{id}/cancel:
    post:
      operationId: bar.orderCancel
      tags:
        - Cafe
      summary: Cancel a cafe order while it is still NEW.
      description: |-
        Refunds the payment if the order was already paid.

        Requirements: FR-BAR-010
      x-roles:
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 3
      x-requirements:
        - FR-BAR-010
      x-tables:
        - bar_orders
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BarOrderDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: ORDER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: INVALID_STATUS_TRANSITION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...